- Updated prime factorization algorithum and came up with better test cases

+++ src/main/java/com/example/coroutine/impl/ParallelProblemsJavaCoRoutineImpl.java
@@ -49,7 +49,7 @@ public class ParallelProblemsJavaCoRoutineImpl implements ParallelProblems {
 
     @NotNull
     @Override
-    public List<Integer> findPrimeFactors(final int n) {
-        return new ArrayList<Integer>();
+    public List<Long> findPrimeFactors(final long n) {
+        return new ArrayList<Long>();
     }
 }
